refactor(geralanding): reorganiza pacotes por etapa em request/response/dto
