+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/GeraLandingCopyExecutionService.java
@@ -1,7 +1,7 @@
 package com.marketinghub.worker.geralanding.copy;
 
 import com.marketinghub.worker.geralanding.copy.GeraLandingExecutionService;
-import com.marketinghub.worker.geralanding.copy.GeraLandingStageExecutionDto;
+import com.marketinghub.worker.geralanding.copy.dto.GeraLandingStageExecutionDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
