+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/dto/GeraLandingStageExecutionDto.java
@@ -1,4 +1,4 @@
-package com.marketinghub.worker.geralanding.copy;
+package com.marketinghub.worker.geralanding.copy.dto;
 
 /** Responsabilidade: representar a execução pendente da etapa de copy no GeraLanding. */
 public record GeraLandingStageExecutionDto(
