+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/deliverables/DeliverablesPendingJobsService.java
@@ -1,5 +1,7 @@
 package com.marketinghub.worker.geralanding.deliverables;
 
+import com.marketinghub.worker.geralanding.deliverables.dto.GeraLandingStageExecutionDetailDto;
+import com.marketinghub.worker.geralanding.deliverables.dto.GeraLandingStageExecutionDeliverablesDto;
 import java.util.List;
 import java.util.Locale;
 import org.springframework.stereotype.Service;
