+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/deliverables/GeraLandingDeliverablesBackendClient.java
@@ -1,5 +1,7 @@
 package com.marketinghub.worker.geralanding.deliverables;
 
+import com.marketinghub.worker.geralanding.deliverables.dto.GeraLandingStageExecutionDetailDto;
+import com.marketinghub.worker.geralanding.deliverables.dto.GeraLandingStageExecutionDeliverablesDto;
 import java.util.List;
 import org.springframework.stereotype.Component;
