+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/deliverables/dto/GeraLandingStageExecutionDeliverablesDto.java
@@ -1,4 +1,4 @@
-package com.marketinghub.worker.geralanding.deliverables;
+package com.marketinghub.worker.geralanding.deliverables.dto;
 
 /** Representa o job pendente da etapa deliverables. */
 public record GeraLandingStageExecutionDeliverablesDto(
@@ -7,7 +7,7 @@ public record GeraLandingStageExecutionDeliverablesDto(
         String stageCode) {
 
     /** Cria um DTO da etapa a partir do DTO base do GeraLanding. */
-    public static GeraLandingStageExecutionDeliverablesDto fromBase(com.marketinghub.worker.geralanding.copy.GeraLandingStageExecutionDto base) {
+    public static GeraLandingStageExecutionDeliverablesDto fromBase(com.marketinghub.worker.geralanding.copy.dto.GeraLandingStageExecutionDto base) {
         return new GeraLandingStageExecutionDeliverablesDto(base.experimentId(), base.idJob(), base.stageCode());
     }
 }
