+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/imageplanning/GeraLandingImagePlanningExecutionService.java
@@ -1,6 +1,7 @@
 package com.marketinghub.worker.geralanding.imageplanning;
 
-import com.marketinghub.worker.geralanding.imageplanning.openai.GeraLandingImagePlanningOpenAiExecutionService;
+import com.marketinghub.worker.geralanding.imageplanning.dto.GeraLandingStageExecutionImagePlanningDto;
+import com.marketinghub.worker.geralanding.imageplanning.request.GeraLandingImagePlanningOpenAiExecutionService;
 import java.util.List;
 import org.springframework.stereotype.Service;
