+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/imageplanning/ImagePlanningPendingJobsService.java
@@ -1,5 +1,7 @@
 package com.marketinghub.worker.geralanding.imageplanning;
 
+import com.marketinghub.worker.geralanding.imageplanning.dto.GeraLandingStageExecutionImagePlanningDto;
+import com.marketinghub.worker.geralanding.imageplanning.dto.GeraLandingStageExecutionDetailDto;
 import java.util.List;
 import java.util.Locale;
 import org.springframework.stereotype.Service;
