+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/presetdesign/PresetDesignPendingJobsService.java
@@ -1,5 +1,7 @@
 package com.marketinghub.worker.geralanding.presetdesign;
 
+import com.marketinghub.worker.geralanding.presetdesign.dto.GeraLandingStageExecutionPresetDesignDto;
+import com.marketinghub.worker.geralanding.presetdesign.dto.GeraLandingStageExecutionDetailDto;
 import java.util.List;
 import java.util.Locale;
 import org.springframework.stereotype.Service;
