+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/wireframe/monitor/WireframePendingJobsService.java
@@ -1,5 +1,7 @@
 package com.marketinghub.worker.geralanding.wireframe.monitor;
 
+import com.marketinghub.worker.geralanding.wireframe.dto.GeraLandingStageExecutionWireframeDto;
+import com.marketinghub.worker.geralanding.wireframe.dto.GeraLandingStageExecutionDetailDto;
 import java.util.List;
 import java.util.Locale;
 import org.slf4j.Logger;
