+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/wireframe/response/RecebeResponse.java
@@ -1,4 +1,4 @@
-package com.marketinghub.worker.geralanding.wireframe.callback;
+package com.marketinghub.worker.geralanding.wireframe.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
